refactor(adapter-api): move the cold-load PR scan into pr_detection::history
Lift scan_history_for_prs out of mainframe-server::chat_deps into the
adapter-neutral module, so the Codex cold-load rescan (a later task) can
call the same function. chat_deps.rs now calls the neutral scan; its
four unit tests moved to mainframe-adapter-api's integration tests, with
their fixture builders copied (not moved) since other chat_deps tests
still use them (todo #339 task 4).
